requirements.txt
# Core scientific stack
pandas==2.2.2
plotly==5.23.0
numpy==1.26.4
matplotlib==3.9.2
seaborn==0.13.2

# Finance data
yfinance==0.2.40

# Web/API requests
requests==2.32.3
newsapi-python==0.2.7

# Machine Learning & NLP basics
scikit-learn==1.5.2
textblob==0.18.0.post0
nltk==3.9.1

# Deep Learning Frameworks
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0

# Hugging Face Transformers + Tokenizers
transformers==4.44.2
datasets==3.0.1
tokenizers==0.19.1

# Explainable AI (XAI)
shap==0.46.0
lime==0.2.0.1

# Experiment tracking & dashboarding
tensorboard==2.17.0
dash==2.17.1
mlflow==2.15.0

# Utilities
tqdm==4.66.5
python-dotenv==1.0.1
